requirements.txt
flask
mutagen
pillow
